LinkedIn/requirements.txt
# LinkedIn AI/ML Auto-Poster Requirements

# Core dependencies
requests>=2.31.0
pyyaml>=6.0.1
feedparser>=6.0.10
beautifulsoup4>=4.12.2
openai>=1.0.0
flask>=3.0.0

# OAuth helper (uses built-in http.server, no extra deps needed)

# Optional: For scheduling
schedule>=1.2.0
python-crontab>=3.0.0

# Optional: For better date parsing
python-dateutil>=2.8.2

